test(bigtable): remove/isolate some test fixtures
